requirements.txt
blinker==1.8.2
cachelib==0.13.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
Flask==3.0.3
Flask-Session==0.8.0
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
msgspec==0.18.6
numpy==2.1.2
oauthlib==3.2.2
packaging==24.1
pandas==2.2.3
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
Werkzeug==3.0.4
